+++ src/negative-hypergeometric-distribution.cpp
@@ -93,9 +93,9 @@ NumericVector cpp_dnhyper(
       Rcpp::warning("NAs introduced by coercion to integer range");
     } else {
       
-      std::vector<double>& tmp = memo[std::make_tuple(i % n.length(),
-                                                      i % m.length(),
-                                                      i % r.length())];
+      std::vector<double>& tmp = memo[std::make_tuple(static_cast<int>(i % n.length()),
+                                                      static_cast<int>(i % m.length()),
+                                                      static_cast<int>(i % r.length()))];
       if (!tmp.size()) {
         tmp = nhyper_table(GETV(n, i), GETV(m, i), GETV(r, i), false);
       }
@@ -159,9 +159,9 @@ NumericVector cpp_pnhyper(
       Rcpp::warning("NAs introduced by coercion to integer range");
     } else {
       
-      std::vector<double>& tmp = memo[std::make_tuple(i % n.length(),
-                                                      i % m.length(),
-                                                      i % r.length())];
+      std::vector<double>& tmp = memo[std::make_tuple(static_cast<int>(i % n.length()),
+                                                      static_cast<int>(i % m.length()),
+                                                      static_cast<int>(i % r.length()))];
       if (!tmp.size()) {
         tmp = nhyper_table(GETV(n, i), GETV(m, i), GETV(r, i), true);
       }
@@ -229,9 +229,9 @@ NumericVector cpp_qnhyper(
       x[i] = NAN;
     } else {
       
-      std::vector<double>& tmp = memo[std::make_tuple(i % n.length(),
-                                                      i % m.length(),
-                                                      i % r.length())];
+      std::vector<double>& tmp = memo[std::make_tuple(static_cast<int>(i % n.length()),
+                                                      static_cast<int>(i % m.length()),
+                                                      static_cast<int>(i % r.length()))];
       if (!tmp.size()) {
         tmp = nhyper_table(GETV(n, i), GETV(m, i), GETV(r, i), true);
       }
@@ -280,9 +280,9 @@ NumericVector cpp_rnhyper(
       x[i] = NA_REAL;
     } else {
       
-      std::vector<double>& tmp = memo[std::make_tuple(i % n.length(),
-                                                      i % m.length(),
-                                                      i % r.length())];
+      std::vector<double>& tmp = memo[std::make_tuple(static_cast<int>(i % n.length()),
+                                                      static_cast<int>(i % m.length()),
+                                                      static_cast<int>(i % r.length()))];
       if (!tmp.size()) {
         tmp = nhyper_table(GETV(n, i), GETV(m, i), GETV(r, i), true);
       }
